rcu rbtree: add redirections for prev/next walk
We need to atomically make the parent/child relationship visible atomically to
the rest of the world. Add redirections in the old nodes to the new nodes which
atomically override the old nodes from the prev/next walk point of view.

Note that it does not affect search, because these do not need to consider the
parents.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>